fix(wots): reorganize block derivation logic and improve mnemonic handling
