+++ packages/babylon-ts-sdk/src/tbv/core/wots/deriveWotsPkHash.ts
@@ -1,8 +1,12 @@
-import { computeWotsPkHash, deriveWotsKeypair, mnemonicToWotsSeed } from "./derivation";
+import {
+  computeWotsBlockPublicKeysHash,
+  deriveWotsBlockPublicKeys,
+  mnemonicToWotsSeed,
+} from "./blockDerivation";
 
 /**
- * Convenience wrapper: derive a WOTS keypair from a mnemonic and return
- * the keccak256 hash of its public key. Handles seed creation and cleanup.
+ * Convenience wrapper: derive WOTS block public keys from a mnemonic and
+ * return the keccak256 hash. Handles seed creation and cleanup.
  *
  * Used before the ETH transaction to produce the `depositorWotsPkHash`
  * that gets committed on-chain.
@@ -15,18 +19,13 @@ export async function deriveWotsPkHash(
 ): Promise<`0x${string}`> {
   const seed = mnemonicToWotsSeed(mnemonic);
   try {
-    const keypair = await deriveWotsKeypair(
+    const publicKeys = await deriveWotsBlockPublicKeys(
       seed,
       peginTxid,
       depositorBtcPubkey,
       appContractAddress,
     );
-    try {
-      return computeWotsPkHash(keypair);
-    } finally {
-      for (const p of keypair.falsePreimages) p.fill(0);
-      for (const p of keypair.truePreimages) p.fill(0);
-    }
+    return computeWotsBlockPublicKeysHash(publicKeys);
   } finally {
     seed.fill(0);
   }
