+++ packages/babylon-ts-sdk/src/tbv/core/wots/index.ts
@@ -1,11 +1,5 @@
-export type { WotsKeypair, WotsPublicKey, WotsKeyProvider } from "./types";
 export {
   mnemonicToWotsSeed,
-  deriveWotsKeypair,
-  keypairToPublicKey,
-  computeWotsPkHash,
-} from "./derivation";
-export {
   deriveWotsBlockPublicKeys,
   computeWotsBlockPublicKeysHash,
 } from "./blockDerivation";
